Add database migration command
